lab2: improve opinions, add votings

+++ protected/data/schema.mysql.sql
@@ -6,7 +6,7 @@ CREATE TABLE product (
     id INTEGER NOT NULL PRIMARY KEY AUTO_INCREMENT,
     name VARCHAR(128) NOT NULL,
     description VARCHAR(256) NOT NULL,
-    mark INTEGER NOT NULL,
+    mark NUMERIC(10, 5) NOT NULL,
     voters INTEGER NOT NULL
 );
